+++ hutool-core/src/main/java/cn/hutool/core/io/resource/ResourceUtil.java
@@ -15,13 +15,15 @@ import cn.hutool.core.util.ClassLoaderUtil;
 
 /**
  * ClassPath资源工具类
+ * 
  * @author Looly
  *
  */
 public class ResourceUtil {
-	
+
 	/**
 	 * 读取Classpath下的资源为字符串，使用UTF-8编码
+	 * 
 	 * @param resource 资源路径，使用相对ClassPath的路径
 	 * @return 资源内容
 	 * @since 3.1.1
@@ -29,9 +31,10 @@ public class ResourceUtil {
 	public static String readUtf8Str(String resource) {
 		return new ClassPathResource(resource).readUtf8Str();
 	}
-	
+
 	/**
 	 * 读取Classpath下的资源为字符串
+	 * 
 	 * @param resource 资源路径，使用相对ClassPath的路径
 	 * @param charset 编码
 	 * @return 资源内容
@@ -40,19 +43,38 @@ public class ResourceUtil {
 	public static String readStr(String resource, Charset charset) {
 		return new ClassPathResource(resource).readStr(charset);
 	}
-	
+
 	/**
 	 * 从ClassPath资源中获取{@link InputStream}
+	 * 
 	 * @param resurce ClassPath资源
 	 * @return {@link InputStream}
+	 * @throws NoResourceException 资源不存在异常
 	 * @since 3.1.2
 	 */
-	public static InputStream getStream(String resurce) {
+	public static InputStream getStream(String resurce) throws NoResourceException {
 		return new ClassPathResource(resurce).getStream();
 	}
-	
+
+	/**
+	 * 从ClassPath资源中获取{@link InputStream}，当资源不存在时返回null
+	 * 
+	 * @param resurce ClassPath资源
+	 * @return {@link InputStream}
+	 * @since 4.0.3
+	 */
+	public static InputStream getStreamSafe(String resurce) {
+		try {
+			return new ClassPathResource(resurce).getStream();
+		} catch (NoResourceException e) {
+			// ignore
+		}
+		return null;
+	}
+
 	/**
 	 * 从ClassPath资源中获取{@link BufferedReader}
+	 * 
 	 * @param resurce ClassPath资源
 	 * @param charset 编码
 	 * @return {@link InputStream}
@@ -61,10 +83,11 @@ public class ResourceUtil {
 	public static BufferedReader getReader(String resurce, Charset charset) {
 		return new ClassPathResource(resurce).getReader(charset);
 	}
-	
+
 	/**
 	 * 获得资源的URL<br>
 	 * 路径用/分隔，例如:
+	 * 
 	 * <pre>
 	 * config/a/db.config
 	 * spring/xml/test.xml
@@ -73,13 +96,14 @@ public class ResourceUtil {
 	 * @param resource 资源（相对Classpath的路径）
 	 * @return 资源URL
 	 */
-	public static URL getResource(String resource) throws IORuntimeException{
+	public static URL getResource(String resource) throws IORuntimeException {
 		return getResource(resource, null);
 	}
-	
+
 	/**
 	 * 获取指定路径下的资源列表<br>
 	 * 路径格式必须为目录格式,用/分隔，例如:
+	 * 
 	 * <pre>
 	 * config/a
 	 * spring/xml
@@ -88,7 +112,7 @@ public class ResourceUtil {
 	 * @param resource 资源路径
 	 * @return 资源列表
 	 */
-	public static List<URL> getResources(String resource){
+	public static List<URL> getResources(String resource) {
 		final Enumeration<URL> resources;
 		try {
 			resources = ClassLoaderUtil.getClassLoader().getResources(resource);
@@ -97,17 +121,18 @@ public class ResourceUtil {
 		}
 		return CollectionUtil.newArrayList(resources);
 	}
-	
+
 	/**
 	 * 获得资源相对路径对应的URL
+	 * 
 	 * @param resource 资源相对路径
 	 * @param baseClass 基准Class，获得的相对路径相对于此Class所在路径，如果为{@code null}则相对ClassPath
 	 * @return {@link URL}
 	 */
-	public static URL getResource(String resource, Class<?> baseClass){
+	public static URL getResource(String resource, Class<?> baseClass) {
 		return (null != baseClass) ? baseClass.getResource(resource) : ClassLoaderUtil.getClassLoader().getResource(resource);
 	}
-	
+
 	/**
 	 * 获取{@link Resource} 资源对象<br>
 	 * 如果提供路径为绝对路径，返回{@link FileResource}，否则返回{@link ClassPathResource}