+++ CHANGELOG.md
@@ -12,9 +12,14 @@
 * 【extra】mail增加抄送和密送支持（感谢【成都】出错）
 * 【poi】ExcelReader别名在返回List时也被支持（第一行）
 * 【poi】ExcelReader增加getSheets和getSheetNames方法（感谢@【帝都】宁静）
+* 【poi】ExcelReader增加readCellValue和readRow方法（感谢@【苏州】马克）
+* 【db】全局数据源工厂独立，使用懒加载方式，消除歧义
+* 【log】全局日志工厂独立，懒加载方式，消除歧义
 
 ### Bug修复
 * 【core】修复获取子路径bug（issue#IHI5K@Gitee）
+* 【poi】修复ExcelReader在读取文件后未关闭导致文件被占用问题（感谢@【昆明】-@_@）
+* 【log】解决Tinylog实现显示类名和行行错误问题
 
 -------------------------------------------------------------------------------------------------------------